requirements.txt
python-docx>=0.8.11
docx2pdf>=0.1.8
pandas>=1.3.0
openpyxl>=3.0.7
fpdf>=1.7.2
babel>=2.9.1
selenium>=4.0.0
webdriver-manager>=3.5.2
tkcalendar>=1.6.1
